+++ HISTORY.rst
@@ -6,3 +6,10 @@ History
 ------------------
 
 * First release on PyPI.
+
+0.5.0 (2020-04-30)
+-------------------
+
+* add command hint, set_server_role
+* export envar
+* add pip source, conda source